requirements.txt
django
gunicorn
django-heroku
requests
django-ckeditor
pillow